Updated Questionnaire
The questionnaire works in that it take information and assigns it to
variables, but I can't get it to assign info to a student class
instance, any ideas? Also do we need to create unique instances, or will
each one be unique as it will have unique info inside, just not sure how
to create unique instances each time in the code.