UC4 Analysis window
This element is effectively complete now. I've created the IO error
exceptions and generally stabilised the whole opening process. The menu
functions all do what they are meant to, and I have added a small help
button for the custom file entry. I've also added a status bar to show
the open file name because it looked cool. ##Needs to launch out of its
folder because of several dependent files## Have a play if you get time.
Try and break it.